requirements.txt
pandas
numpy==1.26.4
scikit-surprise
joblib
streamlit
